+++ pysys-examples/internal/testcases/PySys_internal_097/run.py
@@ -61,3 +61,4 @@ class PySysTest(BaseTest):
 			self.assertGrep('pysys-lang/PySys_NestedTestcase/env.txt', expr='LANG', contains=False)
 		else:
 			self.assertGrep('pysys-lang/PySys_NestedTestcase/env.txt', expr='LANG=my-lang')
+		self.assertGrep('pysys-lang/PySys_NestedTestcase/env.txt', expr='_JAVA_OPTIONS=-Xmx512M')